split storage-kv submodules into files
