+++ storage-kv/src/lib.rs
@@ -1,73 +1,6 @@
-mod kv_store {
-    pub trait KvStore {
-        type InternalError: core::fmt::Debug;
+mod kv_store;
+pub mod mem_kv_store;
 
-        fn get(&self, key: &[u8]) -> Result<Vec<u8>, KvGetError<Self::InternalError>>;
-        fn insert(
-            &mut self,
-            key: Vec<u8>,
-            value: Vec<u8>,
-        ) -> Result<(), KvInsertError<Self::InternalError>>;
-        fn delete(&mut self, key: &[u8]) -> Result<(), KvDeleteError<Self::InternalError>>;
-    }
-
-    #[derive(Debug)]
-    pub enum Infallible {}
-
-    #[derive(Debug)]
-    pub enum KvGetError<InternalError> {
-        NotFound(Vec<u8>),
-        Internal(InternalError),
-    }
-
-    #[derive(Debug)]
-    pub enum KvInsertError<InternalError> {
-        AlreadyExists(Vec<u8>, Vec<u8>),
-        Internal(InternalError),
-    }
-
-    #[derive(Debug)]
-    pub enum KvDeleteError<InternalError> {
-        NotFound(Vec<u8>),
-        Internal(InternalError),
-    }
-}
-
-pub mod mem_kv_store {
-    pub use super::kv_store::*;
-
-    #[derive(Debug, Clone, Default)]
-    pub struct HashMapKv(std::collections::HashMap<Vec<u8>, Vec<u8>>);
-
-    impl KvStore for HashMapKv {
-        type InternalError = Infallible;
-
-        fn get(&self, key: &[u8]) -> Result<Vec<u8>, KvGetError<Infallible>> {
-            self.0
-                .get(key)
-                .cloned()
-                .ok_or(KvGetError::NotFound(key.to_vec()))
-        }
-
-        fn insert(
-            &mut self,
-            key: Vec<u8>,
-            value: Vec<u8>,
-        ) -> Result<(), KvInsertError<Infallible>> {
-            match self.0.insert(key.clone(), value) {
-                Some(old_value) => Err(KvInsertError::AlreadyExists(key, old_value)),
-                None => Ok(()),
-            }
-        }
-
-        fn delete(&mut self, key: &[u8]) -> Result<(), KvDeleteError<Infallible>> {
-            match self.0.remove(key) {
-                Some(_) => Ok(()),
-                None => Err(KvDeleteError::NotFound(key.to_vec())),
-            }
-        }
-    }
-}
 use std::sync::RwLock;
 
 use mem_kv_store::{KvGetError, KvInsertError};
